Fix test suite (Enter watir-webdriver)
* Enter minitest, Remove selenium utilities
* TODO repair other all test (only a part of test_shop.rb works)
* Add README for test and setup
