+++ test/feature/test_ajax_bookmarking.rb
