+++ test/feature/test_communication_views.rb
