+++ test/feature/test_gallery.rb
