+++ test/feature/test_gallery_editing.rb
