+++ test/feature/test_public_views.rb
