+++ test/feature/test_transparent_login_logout.rb
